Refactor classes into respective packages
