+++ src/main/java/aura/io/Parser.java
@@ -1,3 +1,5 @@
+package aura.io;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
