+++ src/main/java/aura/task/Task.java
@@ -1,3 +1,5 @@
+package aura.task;
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
