+++ src/main/java/aura/task/ToDos.java
@@ -1,4 +1,6 @@
-public class ToDos extends Task{
+package aura.task;
+
+public class ToDos extends Task {
     public ToDos(String description) {
         super(description);
     }
