+++ text-ui-test/runtest.bat
@@ -7,10 +7,11 @@ REM delete output from previous run
 if exist ACTUAL.TXT del ACTUAL.TXT
 
 REM compile the code into the bin folder
-javac  -cp ..\src\main\java -Xlint:none -d ..\bin ..\src\main\java\*.java
+javac  -cp ..\src\main\java -Xlint:none -d ..\bin ..\src\main\java\aura\*.java
+
 IF ERRORLEVEL 1 (
     echo ********** BUILD FAILURE **********
-    exit /b 1
+    exit /b 1D
 )
 REM no error here, errorlevel == 0
